backend/requirements.txt
fastapi
uvicorn
pydantic
llama-index
llama-index-llms-openai-like
llama-index-embeddings-openai
llama-index-vector-stores-chroma
llama-index-retrievers-bm25
chromadb
tiktoken
pytest
python-dotenv
